fix: only include commits since previous release in changelog
The previous logic tried to find the tag after the new tag (which doesn't exist yet),
causing PREVIOUS_TAG to be empty and falling back to the last 100 commits from main.
This resulted in too many commit messages being included in the changelog.

Now we simply get the latest existing tag as the previous release, ensuring we only
include commits between the previous release and HEAD.